Fixed Primary keys, no longer arrays, but they are now proper containers.
Fixed portfolio and InvestmentHistoricalMap to use date and extract Ticker to description.

Created inner clase Ticker on PrimaryKey

+++ src/main/java/model/database/containers/Transactions/TransactionValues.java
@@ -4,15 +4,15 @@ import model.database.containers.Inmutable;
 import model.database.containers.Values;
 
 @Inmutable
-public class TransactionValues extends Values<Float> {
+public class TransactionValues extends Values<Double> {
     private static final int NUMBER_OF_FIELDS_IN_TRANSACTION_VALUES = 1;
     private static final int FIELD_AMOUNT_TRADED_POS = 3;
 
-    private TransactionValues(Float[] data) {
+    private TransactionValues(Double[] data) {
         super(data);
     }
 
-    public static TransactionValues getInstance(Float... data) {
+    public static TransactionValues getInstance(Double... data) {
         if (data.length == NUMBER_OF_FIELDS_IN_TRANSACTION_VALUES) {
             return new TransactionValues(data);
         }
@@ -20,10 +20,10 @@ public class TransactionValues extends Values<Float> {
                 + " you introduced: " + data.length);
     }
 
-    public float getAmountTraded(){
+    public double getAmountTraded(){
         return data[FIELD_AMOUNT_TRADED_POS];
     }
-    public float getAbsAmountTraded(){
+    public double getAbsAmountTraded(){
         return StrictMath.abs(data[FIELD_AMOUNT_TRADED_POS]);
     }
